fix encoder specification

+++ su065d4380_interface/include/su065d4380_interface/su065d4380_interface.hpp
@@ -56,7 +56,7 @@ public:
   bool readLastVelocityCommandState() noexcept;
   bool readRightRpm(int16_t &)const noexcept;
   bool readLeftRpm(int16_t &) const noexcept;
-  bool readEncoder(uint16_t &, uint16_t &) const noexcept;
+  bool readEncoder(double &, double &) const noexcept;
   bool readError() const noexcept;
 
   bool writeVelocity(const int16_t &, const int16_t &) noexcept;
